requirements.txt
Flask==3.1.0
pymongo==4.10.1
gunicorn==23.0.0
dnspython==2.7.0
werkzeug==3.1.3
PyJWT==2.9.0
flask-cors==5.0.0
